% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylest2_fit.R
\name{fit_term_usage}
\alias{fit_term_usage}
\title{Internal stylest2 function to compute term usage}
\usage{
fit_term_usage(dfm, smoothing, terms, term_weights)
}
\arguments{
\item{dfm}{a quanteda \code{dfm} object}

\item{smoothing}{the smoothing parameter value for smoothing the dfm. Should 
be a numeric scalar, default to 0.5.}

\item{terms}{If not \code{NULL}, terms to be used in the model. If \code{NULL}, 
use all terms.}

\item{term_weights}{Named vector of distances (or any weights) per term in the 
vocab. Names should correspond to the term.}
}
\value{
A model with with each term that occurs in the text, the frequency of 
use for each author, and the frequency of that terms' occurrence through the texts.
}
\description{
Internal stylest2 function to compute term usage
}
\keyword{internal}
