% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{novels}
\alias{novels}
\title{Excerpts from English novels}
\format{
A dataframe with 21 rows and 3 variables.
}
\source{
Novel excerpts obtained from Project Gutenberg full texts in the
  public domain in the USA. \url{http://gutenberg.org}
}
\usage{
data(novels)
}
\description{
A dataset of text from English novels by Jane Austen, George Eliot, 
and Elizabeth Gaskell.
}
\keyword{datasets}
