% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{novels_dfm}
\alias{novels_dfm}
\title{Novel excerpts in quanteda dfm object}
\format{
A quanteda \code{dfm} with a document variable titled "author".
}
\source{
Novel excerpts obtained from Project Gutenberg full texts in the
  public domain in the USA. \url{http://gutenberg.org}
}
\usage{
data(novels_dfm)
}
\description{
A dataset of text from English novels by Jane Austen, George Eliot, 
and Elizabeth Gaskell. It has been tokenized and processed as a document-feature
matrix in quanteda.
}
\keyword{datasets}
