% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylest2_predict.R
\name{posterior}
\alias{posterior}
\title{Internal stylest2 function to predict posterior likelihoods of authorship.}
\usage{
posterior(log_weights, pred_docs_dfm, speakers, log_prior)
}
\arguments{
\item{log_weights}{The logged author-level term weights from a stylest2 model.}

\item{pred_docs_dfm}{A document-feature matrix for the texts to be predicted.}

\item{speakers}{The speaker labels for \code{pred_docs_dfm} rows.}

\item{log_prior}{A vector of log prior probabilities.}
}
\value{
The posterior likelihoods of authorship for prediction texts.
}
\description{
Generate the posterior likelihood of speaker given their rates and predicted
dfm frequencies.
}
\keyword{internal}
