% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sotu_speeches}
\alias{sotu_speeches}
\title{SOTU speech data}
\format{
A dataframe with 46 rows and 5 variables.
}
\source{
Speeches obtained from the R \code{sotu} package.
}
\description{
A dataset of U.S. presidential State of the Union speeches from 
1974-present
}
\keyword{datasets}
